Rename info.h -> timeinfo.h
