+++ src/libs/antares/benchmarking/i_writer.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include "info.h"
-
 namespace Benchmarking
 {
 class IWriter
